% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eg_data.R
\name{generate_eg_data}
\alias{generate_eg_data}
\title{Generates a test benchmark dataset with varying degrees of wiggliness}
\usage{
generate_eg_data()
}
\value{
data.frame with response column and explanatory variable x column with . It also contains a column truth which is the underlying function without added error.
}
\description{
Generates a test benchmark dataset with varying degrees of wiggliness
}
\examples{
test.data <- generate_eg_data()
plot(test.data$response~test.data$x, pch = 20)
lines(test.data$truth~test.data$x, col = 2, lwd = 1.5)
}
